Add support for repeated classed in pipeline
We introduce the possibility to repeat the same class in a Pipeline.
Previously, the steps were named after the class, which resulted in
the problem that steps had the same name. This is not compatible
with sklearn.Pipeline. The solution is to allow for an optional "name"
key in the class_list. (module:..., class:..., name:...) This way,
the same class can be used twice by providing different names.